requirements.txt
beautifulsoup4==4.12.2
pandas==1.5.0
python-dotenv==1.0.0
requests==2.28.1
selenium==4.9.1
